requirements.txt
selenium==3.141.0
bs4==0.0.1
baidu-aip==2.2.13.0